th06nc: define THGuiCreate & key practice menu / ecl patching funcs; fix misc warnings

+++ thprac_common/src/thprac_games.cpp
@@ -1041,7 +1041,7 @@ void StageWarpsRender(stage_warps_t& warps, std::vector<unsigned int>& out_warp,
 
     switch (warps.type) {
     case stage_warps_t::TYPE_SLIDER:
-        ImGui::SliderInt(warps.label, (int*)&out_warp[level], 0, warps.section_param.size() - 1, warps.section_param[out_warp[level]].label);
+        ImGui::SliderInt(warps.label, (int*)&out_warp[level], 0, (int)warps.section_param.size() - 1, warps.section_param[out_warp[level]].label);
         break;
     case stage_warps_t::TYPE_COMBO:
         if (ImGui::BeginCombo(warps.label, warps.section_param[out_warp[level]].label)) {
@@ -1075,7 +1075,7 @@ void StageWarpsRender(stage_warps_t& warps, std::vector<unsigned int>& out_warp,
     }
 
     if (warps.section_param[out_warp[level]].phases) {
-        ImGui::PushID(level + 1);
+        ImGui::PushID((int)level + 1);
         StageWarpsRender(*warps.section_param[out_warp[level]].phases, out_warp, level + 1);
         ImGui::PopID();
     }
